renamed StandardFAQ to PinnedFAQ, renamed faq_content props to question_answer_pairs
