+++ client/src/components/PinnedFAQ/PinnedFAQ.stories.tsx
@@ -5,29 +5,29 @@ import common_lang from "src/panels/panel_declarations/misc/key_concept_panels/c
 
 import { ComponentProps } from "src/types/util_types.d";
 
-import { faq_content_maker } from "./faq_utils";
-import { StandardFAQ } from "./StandardFAQ";
+import { qa_pairs_maker } from "./faq_utils";
+import { PinnedFAQ } from "./PinnedFAQ";
 
 import common_subapp_lang from "./common_faq_questions.yaml";
-import sample_lang from "./StandardFAQ.yaml";
+import sample_lang from "./PinnedFAQ.yaml";
 
 export default {
-  title: "StandardFAQ",
-  component: StandardFAQ,
+  title: "PinnedFAQ",
+  component: PinnedFAQ,
 } as Meta;
 
-type StandardFAQProps = ComponentProps<typeof StandardFAQ>;
+type PinnedFAQProps = ComponentProps<typeof PinnedFAQ>;
 
-const Template: Story<StandardFAQProps> = (args) => <StandardFAQ {...args} />;
+const Template: Story<PinnedFAQProps> = (args) => <PinnedFAQ {...args} />;
 
 const q_a_keys = ["question1", "question2", "question3"];
 
-const faq_content = faq_content_maker(
+const question_answer_pairs = qa_pairs_maker(
   { bundles: [sample_lang, common_lang, common_subapp_lang] },
   q_a_keys
 );
 
 export const Basic = Template.bind({});
 Basic.args = {
-  faq_content: faq_content,
+  question_answer_pairs: question_answer_pairs,
 };
