+++ client/src/components/PinnedFAQ/PinnedFAQ.yaml
@@ -1,4 +1,4 @@
-# FAQ keys for the StandardFAQ story
+# FAQ keys for the PinnedFAQ story
 
 question1_q:
   en: |
