+++ client/src/components/PinnedFAQ/igoc_explorer_questions.yaml
